+++ src/middleware/AuthenticateToken.ts
@@ -15,7 +15,7 @@ export const authenticateToken = async (
 ) => {
   const authHeader = req.headers["authorization"];
   const token = authHeader && authHeader.split(" ")[1];
-  if (!token) return res.status(401).json({ message: { token: token } });
+  if (!token) return res.status(401).json({ message: "Unauthorized" });
 
   try {
     const decoded = jwt.verify(
