+++ src/usecase/LoginUser.usecase.ts
@@ -1,7 +1,7 @@
-import { LoginDto } from "../../core/dto/Login.dto";
-import { ErrorResponse } from "../../core/entity/ErrorRespose.entity";
-import { LoggedDataEntity } from "../../core/entity/LoggedData.entity";
-import { AuthenticationRepository } from "../../core/repository/AuthenticationRepository";
+import { LoginDto } from "../core/dto/Login.dto";
+import { ErrorResponse } from "../core/entity/ErrorRespose.entity";
+import { LoggedDataEntity } from "../core/entity/LoggedData.entity";
+import { AuthenticationRepository } from "../core/repository/AuthenticationRepository";
 
 export class LoginUserUsecase {
   constructor(private authenticationRepository: AuthenticationRepository) {}
